added users router
