+++ projojo_backend/routes/test_router.py
@@ -12,23 +12,6 @@ from domain.models import BusinessProjects, Skill
 
 router = APIRouter(prefix="/test", tags=["Test Endpoints"])
 
-
-@router.get("/users")
-async def get_all_users():
-    """
-    Get all users for debugging purposes
-    """
-    users = user_repo.get_all()
-    return users
-
-@router.get("/users/{email}")
-async def get_user(email: str = Path(..., description="User email")):
-    """
-    Get a specific user by email
-    """
-    user = user_repo.get_by_id(email)
-    return user
-
 @router.post("/upload", status_code=201)
 async def upload_file(file: UploadFile = File(...)):
     """
